Serialie BigInt to float
The error is that SyncProgressJob is receiving BigDecimal values for progress and duration, but Sidekiq requires native JSON types. The params[:progress].to_d and params[:duration].to_d return BigDecimal objects, which can't be serialized to JSON by Sidekiq.

+++ app/controllers/player_controller.rb
@@ -84,8 +84,8 @@ class PlayerController < ApplicationController
 
     content_id = params[:id]
     episode_id = params[:episode_id].presence
-    progress = params[:progress].to_d
-    duration = params[:duration].to_d
+    progress = params[:progress].to_f
+    duration = params[:duration].to_f
 
     # 1. Guardar solo en Redis
     episode_key = episode_id.presence || 'movie'
